% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.r
\name{default}
\alias{default}
\alias{default<-}
\title{Gets and sets the default value of a dictionary.}
\usage{
default(dict)

default(dict) <- value
}
\arguments{
\item{dict}{an existing dictionary object}

\item{value}{the new default value}
}
\value{
the modified dictionary
}
\description{
Functions to get and set the default value of a dictionary. The default value is returned
when a non-existing key is accessed.
}

