% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{dict_operators}
\alias{dict_operators}
\alias{[[.dict}
\alias{$.dict}
\alias{$<-.dict}
\alias{[[<-.dict}
\alias{[.dict}
\alias{[<-.dict}
\title{Access or replace values of a dictionary}
\usage{
\method{[[}{dict}(dict, key)

\method{$}{dict}(dict, key)

\method{$}{dict}(dict, key) <- value

\method{[[}{dict}(dict, key) <- value

\method{[}{dict}(dict, ...)

\method{[}{dict}(dict, ...) <- value
}
\arguments{
\item{dict}{object from which to access value(s) or in which to replace value(s)}

\item{key}{index or indices specifying the value to access or replace}

\item{value}{list or vector of value(s) to replace}

\item{...}{additional keys}
}
\description{
Access or replace one or multiple values of a dictionary using keys.
}
\details{
These operators work equivalently to the list operators \code{[}, \code{[[} and \code{$}.
These operators do not partially match keys. Additionally, assigning \code{NULL} to one
of the keys does not remove the entry from the dictionary; instead, it sets the value
corresponding to that key to \code{NULL} (use \link{omit} to remove entries).
}
