% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.r
\name{entries}
\alias{entries}
\title{Converts a dictionary to a list of key/value pairs}
\usage{
entries(dict)
}
\arguments{
\item{dict}{a dictionary}
}
\value{
a list containing lists with two items, \code{key} and \code{value}, for each entry in the
dictionary.
}
\description{
Converts a dictionary to a list of key/value pairs
}
\examples{
distances <- dict(Mercury = 0.387, Venus = 0.723, Earth = 1)
for (e in entries(distances))
   cat('The distance between planet', e$key, ' and the Sun is', e$value, ' AU.\\n')
}

