% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\name{entry}
\alias{entry}
\title{Returns a list with elements \code{key} and \code{value}.}
\usage{
entry(key, value)
}
\arguments{
\item{key}{the key}

\item{value}{the value}
}
\description{
Returns a list with elements \code{key} and \code{value}.
}
