% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R, R/utils.R
\name{defaults}
\alias{defaults}
\alias{extend}
\title{Merge two dictionaries or named lists.}
\usage{
defaults(x, defaults)

extend(x, ...)

defaults(x, defaults)
}
\arguments{
\item{x}{original object; a dictionary or named list}

\item{defaults}{dictionary or named list providing default values for \code{x}}

\item{...}{dictionaries or named lists that override \code{x}. Named parameters
can also be used.}
}
\description{
Merge two dictionaries or named lists.
}
\details{
\code{c} or \code{extend} overrides the entries in the first dictionary with
entries from the following objects, left to right (each object overrides
entries from the previous argument).

\code{defaults} augments the entries in the first dictionary with entries from the second dictionary
that were not present in the first.
}
\examples{
x <- dict(a=1, b=2, c=3)
y <- dict(a="a", b="b")

c(x, y) # returns dict(a="a", b="b", c=3)
c(x, c=9, d=16) # returns dict(a=1, b=2, c=9, d=16)
defaults(x, dict(a=10, d=10)) # returns dict(a=1, b=2, c=3, d=10)
}
