% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has}
\alias{has}
\title{Check whether a dictionary contains the specified key(s)}
\usage{
has(dict, keys)
}
\arguments{
\item{dict}{a dictionary}

\item{keys}{a character vector of keys}
}
\description{
Check whether a dictionary contains the specified key(s)
}
