% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash.r
\name{hash_code}
\alias{hash_code}
\title{Computes the hash code of an arbitrary object.}
\usage{
hash_code(x)
}
\arguments{
\item{x}{object for which to compute the hash}
}
\value{
an integer value
}
\description{
Computes the hash code of an object. The default implementation serializes the object and computes the
hash using the \link[digest]{digest} function in the digest package (algorithm xxhash32).
}
\details{
hash_code is a generic function, so its default behavior can be overridden for a class.
}

