% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immutable.R
\name{immutable}
\alias{immutable}
\title{Transforms any collection into an immutable collection.}
\usage{
immutable(coll)
}
\arguments{
\item{coll}{collection to be made immutable}
}
\value{
a copy of the collection, marked as immutable
}
\description{
Returns a copy of the collection that cannot be modified using the bracket, double bracket
or dollar operators.
}
