% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{invert}
\alias{invert}
\title{Inverts a dictionary}
\usage{
invert(dict)
}
\arguments{
\item{dict}{a dictionary}
}
\value{
a new dictionary with keys and values swapped
}
\description{
Returns a dictionary where keys and values are swapped. Values are coerced to character
keys, if needed.
}
