% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immutable.R
\name{is_immutable}
\alias{is_immutable}
\title{Checks whether a collection is immutable}
\usage{
is_immutable(coll)
}
\arguments{
\item{coll}{collection}
}
\value{
TRUE if immutable, FALSE otherwise
}
\description{
Checks whether a collection is immutable
}
