% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.r
\name{keys}
\alias{keys}
\alias{keys<-}
\title{Returns or assigns the keys of the provided dictionary.}
\usage{
keys(dict)

keys() <- value
}
\arguments{
\item{dict}{a dictionary}
}
\value{
a character vector containing the keys of the dictionary
}
\description{
Returns or assigns the keys of the provided dictionary.
}

