% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\name{length.dict}
\alias{length.dict}
\title{Get the length of the dictionary}
\usage{
\method{length}{dict}(x)
}
\arguments{
\item{x}{a dictionary object}
}
\value{
the number of unique keys stored in the dictionary.
}
\description{
Returns the number of unique keys in the dictionary.
}
