% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.r
\name{map_dict}
\alias{map_dict}
\title{Apply a function to each entry of the dictionary}
\usage{
map_dict(dict, .f, ...)
}
\arguments{
\item{dict}{A dictionary}

\item{.f}{A function, formula or character vector.
If a function, it is called with two arguments, the key and the value.
If a formula, .x corresponds to the key and .y to the value.
If a character vector, it will return the keys specified in the vector}
}
\value{
A dictionary with the same keys as \code{dict}, and values given by the return value of \code{.f}.
}
\description{
Calls a function on each entry and builds a dictionary from the transformed input. See \link[purrr]{map}.
}

