% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.r
\name{omit}
\alias{omit}
\title{Removes key(s) from a dictionary or named list}
\usage{
omit(dict, ...)
}
\arguments{
\item{dict}{dictionary or named list}

\item{...}{character vector(s) of keys to remove}
}
\value{
a new object without the entries corresponding to the specified keys
}
\description{
Returns a new dictionary or named list with the specified keys omitted.
}

