% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.r
\name{print.dict}
\alias{print.dict}
\title{Prints the contents of a dictionary}
\usage{
\method{print}{dict}(x, digits = NULL)
}
\arguments{
\item{x}{dictionary}

\item{digits}{minimal number of significant digits}
}
\description{
Prints the contents of a dictionary
}

