% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\name{str.dict}
\alias{str.dict}
\title{Display structure of dictionary}
\usage{
\method{str}{dict}(dict, ...)
}
\arguments{
\item{dict}{a dictionary}

\item{...}{additional parameters passed to \link[utils]{str}}
}
\description{
See \link[utils]{str}.
}
