% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\name{values}
\alias{values}
\alias{values<-}
\title{Returns or assigns the values of the provided dictionary.}
\usage{
values(dict)

values(dict, values) <- value
}
\arguments{
\item{dict}{a dictionary}

\item{value}{new values}
}
\value{
a list containing the values of the dictionary, or the updated
  dictionary
}
\description{
Returns or assigns the values of the provided dictionary.
}
\examples{
map <- dict(a=1, b=2)

values(map) <- list("A", "B") # now map == dict(a="A", b="B")
}
