% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\name{as.list.dict}
\alias{as.list.dict}
\title{Coerces the dictionary to a named list}
\usage{
\method{as.list}{dict}(dict)
}
\arguments{
\item{dict}{dictionary}
}
\value{
a named list with the same keys and values
}
\description{
Coerces the dictionary to a named list
}
