% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.r
\name{as_dict}
\alias{as_dict}
\title{Coerces an object to a dictionary}
\usage{
as_dict(obj)
}
\arguments{
\item{obj}{a named list or vector.}
}
\value{
a dictionary containing the same keys and values as the input object.
}
\description{
Coerces a named list or vector to a dictionary.
}

