% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.r
\name{entries}
\alias{entries}
\title{Converts a dictionary to a list of key/value pairs}
\usage{
entries(dict)
}
\arguments{
\item{dict}{a dictionary}
}
\value{
a list containing lists with two items, \code{key} and \code{value}, for each entry in the
dictionary.
}
\description{
Converts a dictionary into a list of entries containing key/value pairs, of the form
\code{list(entry(key = key1, value = value1), entry(key = key2, value = value2), ...)}.
Each element of the list is a list containing a key element and a value element, created using
the \link{entry} function. A list of entries can be collected back into a dictionary using
the \link{collect} function.
}
\examples{
solar_system <- dict(Mercury = 0.387, Venus = 0.723, Earth = 1, Mars = 1.524)
for (e in entries(solar_system))
   cat('The distance between planet', e$key, ' and the Sun is', e$value, ' AU.\\n')

inner_solar_system <- entries(solar_system) \%>\%
             keep(function(e) e$value <= 1) \%>\%
             collect()
}

