% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.r
\name{extend}
\alias{defaults}
\alias{extend}
\title{Merge two dictionaries or named lists.}
\usage{
extend(x, ...)

defaults(x, defaults)
}
\arguments{
\item{x}{original object; a dictionary or named list}

\item{...}{dictionaries or named lists that override \code{x}}

\item{defaults}{dictionary or named list providing default values for \code{x}}
}
\description{
Merge two dictionaries or named lists.
}
\details{
\code{extend} overrides the entries in the first dictionary with entries from the following objects,
left to right (each object overrides entries from the previous argument).

\code{defaults} augments the entries in the first dictionary with entries from the second dictionary
that were not present in the first.
}

