% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immutable.R
\name{immutable}
\alias{immutable}
\alias{mutable}
\title{Transforms any collection into an immutable collection.}
\usage{
immutable(coll)

mutable(coll)
}
\arguments{
\item{coll}{collection to be made immutable}
}
\value{
A copy of the collection, marked as immutable. If \code{mutable} is
  called, then a copy of the collection with the immutability flag removed is
  returned.
}
\description{
Returns a copy of the collection that cannot be modified using the bracket, double bracket
or dollar operators.
}
\details{
\code{mutable} will make an immutable collection mutable again.
}
\examples{
a <- list(1, 2, 3)
a[[1]] <- 2  # ok

a <- immutable(a)
# a[[1]] <- 2  will throw an error
}
