% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\name{keys}
\alias{keys}
\alias{keys<-}
\title{Returns or assigns the keys of the provided dictionary.}
\usage{
keys(dict)

keys(dict) <- value
}
\arguments{
\item{dict}{a dictionary}

\item{value}{new keys to assign}
}
\value{
a character vector containing the keys of the dictionary
}
\description{
Returns or assigns the keys of the provided dictionary.
}
