% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\name{make_dict}
\alias{make_dict}
\title{Create a new dictionary from a list or vector of keys and values}
\usage{
make_dict(keys, values, default = NULL, strict = FALSE)
}
\arguments{
\item{keys}{a vector of character keys}

\item{values}{a vector or list of values to be associated to the corresponding keys.
Alternatively, a purrr-style mapper function can be passed, which will be
called with each key as a parameter.}

\item{default}{a default value for the dictionary (see \link{default_dict})}

\item{strict}{whether the dictionary should be strict and raise an exception when a non-existing key is accessed (see \link{strict_dict})}
}
\value{
a new dictionary
}
\description{
Create a new dictionary from separate keys and values.
}
\examples{
# create an employee dict out of keys and values
employee <- make_dict(keys=c('name', 'age'), values=list('John', 60))

# create a dict by applying a function to the keys
letter_to_LETTER <- make_dict(letters, toupper)
}
