% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrr_ext.R
\name{map_dict}
\alias{map_dict}
\alias{keep_dict}
\alias{discard_dict}
\alias{compact_dict}
\title{Map, filter, compact on keys and values of the dictionary}
\usage{
map_dict(x, .f, ...)

keep_dict(x, .p, ...)

discard_dict(x, .p, ...)

compact_dict(x)
}
\arguments{
\item{x}{A dictionary or named list}

\item{.f}{A function, formula or character vector.}

\item{...}{Optional parameters to pass to \code{.f} or \code{.p}}

\item{.p}{A predicate function returning a logical value.}
}
\value{
For \code{map_dict}, a dictionary with the same keys as \code{dict}, and values
given by the return value of \code{.f}. For \code{keep_dict}, \code{discard_dict}, and
\code{compact_dict}, a dictionary containing the entries that passed the filter.
}
\description{
\code{map_dict} calls a function on each (key, value) pair.
(see \link[purrr]{map} in package \code{purrr}).
}
\details{
\code{keep_dict} and \code{discard_dict} keep and discard elements based on the return value of the
predicate function (see \link[purrr]{keep} and \link[purrr]{discard} in package \code{purrr}).

\code{compact_dict} returns a dictionary with all the \code{NULL} values removed.

Functions are called with arguments (key, value). If the function or predicate
argument is a mapper function, \code{.x} will correspond to the key and
\code{.y} to the value.
}
