% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\name{print.dict}
\alias{print.dict}
\title{Prints the contents of a dictionary}
\usage{
\method{print}{dict}(x, ...)
}
\arguments{
\item{x}{dictionary}

\item{...}{other parameters forwarded to print()}
}
\description{
Prints the contents of a dictionary
}
