% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.r
\name{values}
\alias{values}
\alias{values<-}
\title{Returns or assigns the values of the provided dictionary.}
\usage{
values(dict)

values(dict) <- value
}
\arguments{
\item{dict}{a dictionary}
}
\value{
a list containing the values of the dictionary
}
\description{
Returns or assigns the values of the provided dictionary.
}

